requirements.txt
requests==2.31.0
pydantic==2.5.0
python-dotenv==1.0.0
tqdm==4.66.1
PyYAML==6.0.1
